Add Ionic router setup and home page
