+++ src/theme/variables.css
@@ -1,0 +1,6 @@
+:root {
+  --ion-background-color: #ffffff;
+  --ion-text-color: #000000;
+  --ion-color-primary: #3880ff;
+  --ion-color-primary-contrast: #ffffff;
+}
